backend/pyproject.toml
[tool.poetry]
name = "Taskera-Ai-Project"
version = "1.0.0"
description = "Multi-functional AI agent with ACP/MCP architecture."
authors = ["Mubashir Hassan <mubashir.hassan@devisgon.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
acp-sdk = "^0.4.0"
langchain-google-genai = "^3.0.3"
httpx = "^0.28.1"

# --- API & Server Dependencies ---
fastapi = "^0.115.0"
uvicorn = ">=0.29.0"

# --- Core Tools ---
apscheduler = "^3.10.4"
supabase = "^2.5.2"
chromadb = "^1.3.4"
langchain-community = "^0.3.0"
langchain-text-splitters = "^0.3.0"
langchain = "^1.0.0"

# --- Document Processing ---
pypdf = "^4.3.1"
docx2txt = "^0.8"
unstructured = "^0.18.15"
# Note: unstructured-inference is removed to save space/GPU reqs
pytesseract = "^0.3.10"
pillow = "^12.0.0"
dateparser = "^1.2.0"
playwright = "^1.45.1"

# --- Search Tools ---
# You requested both; keeping them ensures LangChain finds what it needs
ddgs = "*" 
duckduckgo-search = "^5.3.1" 
pyowm = "^3.3.0"
wikipedia = "^1.4.0"
google-api-python-client = "^2.187.0"
google-auth-httplib2 = "^0.2.1"
google-auth-oauthlib = "^1.2.3"
langchain-chroma = "^1.0.0"

# --- PDF & Image Utils ---
pdfminer-six = "^20240706"
pdf2image = "^1.17.0"
pi-heif = "^1.1.1"
opencv-python-headless = "^4.10.0"
matplotlib = "^3.9.0"

# --- Security & Math ---
slowapi = "^0.1.9"
starlette-csrf = "^3.0.0"
itsdangerous = "^2.2.0"
numexpr = "^2.10.0"
langchain-core = "^1.1.0"
aiosqlite = "^0.21.0"
langgraph-checkpoint-sqlite = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"